refactor: Move files into mobile duck
