+++ src/ducks/mobile/utils.js
